fixes #67
