+++ frontend/rievent-frontend/src/components/Drawer/Main.js
@@ -11,9 +11,9 @@ export const Main = styled('main', { shouldForwardProp: (prop) => prop !== 'open
         }),
         marginLeft: `-${drawerWidth}px`,
         ...(open && {
-            transition: theme.transitions.create('margin', {
-            easing: theme.transitions.easing.easeOut,
-            duration: theme.transitions.duration.enteringScreen,
+                transition: theme.transitions.create('margin', {
+                easing: theme.transitions.easing.easeOut,
+                duration: theme.transitions.duration.enteringScreen,
             }),
             marginLeft: 0,
         }),
